admin/requirements.txt
django==5.2.7
python-dotenv==1.0.0
chromadb==1.4.0
langchain-text-splitters==1.1.0
